ARM: BCM2835: Add a Raspberry Pi-specific clock driver.
Unfortunately, the clock manager's registers are not accessible by the
ARM, so we have to request that the firmware please control our clocks
for us.

Not all of the clocks here get used by their consumers, like EMMC and
UART.  We probably want to connect those up at some point, so that
more of our detection is automatic instead of hardcoded in DT.  Until
then, we mark them as CLK_IGNORE_UNUSED so they don't get turned off
while they're still needed.

Signed-off-by: Eric Anholt <eric@anholt.net>