cpufreq: add raspberry cpufreq driver
The CPU clock is managed by the VC core. Drivers to talk with the
the VC firmware and to set/get all SoC clocks (including CPU clock)
are already there.

This driver implemets cpufreq set/get operations on the top of the
existing clk driver and it uses directly the firmware driver for
initial max/min rate configuration.

Signed-off-by: Andrea Merello <andrea.merello@gmail.com>